jobs/requirements.txt
asgiref==3.4.1
Babel==2.9.1
Django==4.0.1
django-countries==7.2.1
django-money==2.1.1
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2==2.9.3
py-moneyed==1.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
